\name{Hyperedge}
\alias{Hyperedge}
\title{Constructor for Hyeredge objects}
\description{
  A convenience constructor for \code{\link{Hyperedge-class}} objects
}
\usage{
Hyperedge(nodes, label = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nodes}{Character vector of nodes that are part of the hyperedge}
  \item{label}{A character string describing the hyperedge}
}
\value{
  An object of class \code{\link{Hyperedge-class}}
}
\author{Seth Falcon}
\seealso{
  \code{\link{Hyperedge-class}}
  \code{\link{Hypergraph-class}}
}
\keyword{classes}
