\name{Hypergraph}
\alias{Hypergraph}
\title{Constructor for Hypergraph objects}
\description{
  A convenience constructor for \code{link{Hypergraph-class}} objects
}
\usage{
Hypergraph(nodes, hyperedges)
}
\arguments{
  \item{nodes}{A vector of nodes (character)}
  \item{hyperedges}{A list of \code{\link{Hyperedge-class}} objects}
}
\value{
  An object of class \code{\link{Hypergraph-class}}
}
\author{Seth Falcon}
\seealso{
  \code{\link{Hypergraph-class}}
  \code{\link{Hyperedge-class}}
  \code{\link{DirectedHyperedge-class}}
  }
\keyword{classes}
